+++ libs/cdk/zone-less/rxjs/src/internals/intervalProvider.ts
@@ -1,0 +1,30 @@
+import { setInterval } from '@rx-angular/cdk/zone-less/browser';
+
+type SetIntervalFunction = (
+  handler: () => void,
+  timeout?: number,
+  ...args: any[]
+) => number;
+type ClearIntervalFunction = (handle: number) => void;
+
+interface IntervalProvider {
+  setInterval: SetIntervalFunction;
+  clearInterval: ClearIntervalFunction;
+  delegate?: Omit<IntervalProvider, 'delegate'>;
+}
+
+export const intervalProvider: IntervalProvider = {
+  // When accessing the delegate, use the variable rather than `this` so that
+  // the functions can be called without being bound to the provider.
+  setInterval(handler: () => void, timeout?: number, ...args) {
+    const { delegate } = intervalProvider;
+    if (delegate?.setInterval) {
+      return delegate.setInterval(handler, timeout, ...args);
+    }
+    return setInterval(handler, timeout);
+  },
+  clearInterval(handle) {
+    const { delegate } = intervalProvider;
+    return (delegate?.clearInterval || clearInterval)(handle as any);
+  },
+};