+++ libs/state/selections/src/lib/interfaces/non-undefined.ts
@@ -1,0 +1,1 @@
+export type NonUndefined<T> = T extends undefined ? never : T;