+++ libs/template/experimental/if/src/lib/if.directive.ts
@@ -9,30 +9,30 @@ import {
   OnInit,
   SimpleChanges,
   TemplateRef,
-  ViewContainerRef
+  ViewContainerRef,
 } from '@angular/core';
 import { coerceAllFactory, coerceObservable } from '@rx-angular/cdk/coercing';
 import {
   createTemplateNotifier,
-  RxNotificationKind
+  RxNotificationKind,
 } from '@rx-angular/cdk/notifications';
 import { RxStrategyProvider } from '@rx-angular/cdk/render-strategies';
 import {
   createTemplateManager,
-  RxTemplateManager
+  RxTemplateManager,
 } from '@rx-angular/cdk/template';
 import {
   NextObserver,
   Observable,
   ReplaySubject,
   Subject,
-  Subscription
+  Subscription,
 } from 'rxjs';
 import { mergeAll } from 'rxjs/operators';
 import {
   RxIfTemplateNames,
   rxIfTemplateNames,
-  RxIfViewContext
+  RxIfViewContext,
 } from './model/index';
 
 @Directive({
@@ -53,7 +53,6 @@ export class RxIf<U> implements OnInit, OnDestroy, OnChanges {
     this.observablesHandler.next(coerceObservable(potentialObservable));
   }
 
-  /* eslint-disable @angular-eslint/no-input-rename */
   @Input('rxIfStrategy')
   set strategy(strategyName: Observable<string> | string | null | undefined) {
     this.strategyHandler.next(strategyName);