+++ libs/template/let/src/lib/let.directive.ts
@@ -29,6 +29,7 @@ import {
   RxTemplateManager,
   RxViewContext
 } from '@rx-angular/cdk/template';
+
 import {
   defer,
   merge,
@@ -298,7 +299,6 @@ export class LetDirective<U> implements OnInit, OnDestroy, OnChanges {
     this._renderObserver = callback;
   }
 
-  // eslint-disable-next-line @angular-eslint/no-input-rename
   @Input('rxLetParent') renderParent = this.strategyProvider.config.parent;
 
   @Input('rxLetPatchZone') patchZone = this.strategyProvider.config.patchZone;